requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0

beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==4.2.17
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.14.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
numpy==1.24.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3

psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==3.5.3
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
